+++ crates/nostr-relay-pool/src/prelude.rs
@@ -21,4 +21,5 @@ pub use crate::pool::constants::*;
 pub use crate::pool::options::*;
 pub use crate::pool::{self, *};
 pub use crate::relay::{self, *};
+pub use crate::stream::{self, *};
 pub use crate::*;