+++ crates/nostr-relay-pool/src/stream.rs
@@ -2,11 +2,20 @@
 // Copyright (c) 2023-2025 Rust Nostr Developers
 // Distributed under the MIT software license
 
+//! Stream
+
 use std::pin::Pin;
 use std::task::{Context, Poll};
 
 use async_utility::futures_util::Stream;
 use tokio::sync::mpsc::Receiver;
+
+/// Boxed stream
+#[cfg(not(target_arch = "wasm32"))]
+pub type BoxedStream<T> = Pin<Box<dyn Stream<Item = T> + Send>>;
+/// Boxed stream
+#[cfg(target_arch = "wasm32")]
+pub type BoxedStream<T> = Pin<Box<dyn Stream<Item = T>>>;
 
 #[derive(Debug)]
 pub(crate) struct ReceiverStream<T> {